requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
brewer2mpl==1.4.1
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
checksumdir==1.1.6
chicken-turtle-project==2.3.0
chicken-turtle-util==1.0.0
click==8.0.1
collections-extended==1.0.2
colorama==0.4.1
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.0
dohq-tfs==1.0.175
first==2.0.2
Flask==2.0.1
flask-ldap3-login==0.9.16
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
gitdb2==2.0.5
GitPython==2.1.11
greenlet==1.1.1
idna==2.10
importlib-metadata==0.18
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.1.0
ldap3==2.9.1
MarkupSafe==2.0.1
matplotlib==3.1.1
more-itertools==7.1.0
networkx==2.3
nltk==3.6.2
ntlm-auth==1.5.0
numpy==1.18.4
openrefine-client==0.3.8
packaging==19.0
palettable==3.3.0
pandas==0.24.2
pdfminer==20191125
pdfminer.six==20181108
pdfminer.six-i==20190823
pdfplumber==0.5.14
pdfplumber-i==0.6.4
Pillow==6.2.0
pip-tools==1.6.5
pluggy==0.12.0
plumbum==1.6.7
py==1.8.0
pyad==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.45
pycparser==2.20
pycryptodome==3.9.7
PyGetWindow==0.0.6
PyMsgBox==1.0.7
pyodbc==4.0.27
pypandoc==1.4
pyparsing==2.4.5
pyPdf==1.13
PyPDF4==1.27.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.21
pytest==5.0.1
python-dateutil==2.8.1
python-Levenshtein-wheels==0.13.2
PyTweening==1.0.3
pytz==2019.1
pywin32==228
regex==2021.8.28
requests==2.25.1
requests-ntlm==1.1.0
selenium==3.141.0
six==1.13.0
smmap2==2.0.5
sortedcontainers==2.1.0
SQLAlchemy==1.4.22
tqdm==4.62.2
unicodecsv==0.14.1
urllib2-file==0.2.1
urllib3==1.25.3
Versio==0.4.0
Wand==0.5.7
wcwidth==0.1.7
Werkzeug==2.0.1
WTForms==2.3.3
xlrd==1.2.0
zipp==0.5.2
